alteraçoes de tipos

+++ src/database/migrations/20200415191903_create_market.js
@@ -1,16 +1,15 @@
-
-exports.up = function(knex) {
-    return knex.schema.createTable('market', function(table){
-        table.increments();
-        table.string('name').notNullable();
-        table.string('street').notNullable();
-        table.string('district').notNullable();
-        table.string('phone');
-        table.decimal('flagCountAny');
-        table.decimal('flagCountMonth');
-    });
+exports.up = function (knex) {
+  return knex.schema.createTable("market", function (table) {
+    table.increments();
+    table.string("name").notNullable();
+    table.string("street").notNullable();
+    table.string("district").notNullable();
+    table.string("phone");
+    table.integer("flagCountAny");
+    table.integer("flagCountMonth");
+  });
 };
 
-exports.down = function(knex) {
-  return knex.schema.dropTable('market');
+exports.down = function (knex) {
+  return knex.schema.dropTable("market");
 };
